requirements.txt
spotipy
mysql-connector-python
python-dotenv
